02-py-async-jobs/requirements.txt
jetpack-io>=0.5.1-dev20210921
flask
fastapi
flask[async]
boto3
uvicorn